+++ changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MonadLog instance for `MockchainT`
 - API documentation published on github pages (https://j-mueller.github.io/sc-tools/)
 - A mockchain test that shows how to increase the maximum tx size
+- `Convex.MockChain.Utils.runMockchainPropWith` to run a mockchain action in a quickcheck property with custom node params and initial distribution
 
 ### Deleted